Initial commit
Start of the tool-versions-update-action, a GitHub Action intended for
keeping tools defined in a `.tool-versions` file up-to-date. As a start
this is a composite action pulling together a couple of actions combined
with a custom Bash script that add the updating functionality.

In this initial versions, this action checks out your repository, then
install asdf (used to do the updates) [1], then updates tools defined in
the `.tool-versions` file, and finally creates a Pull Request for the
updates.

It can currently be configured to update up to a maximum number of tools
which can help reduce the scope of created Pull Requests if that's
desired. This behavior is inspired by Dependabot's [2] open Pull Request
limit option.

Other than that, this initial release sets up some tooling with configs.
EditorConfig [3] for collaboration, hadolint [4] for Dockerfile linting,
ShellCheck [5] for Shell script linting/vetting, Yamllint [6] for y(a)ml
file linting, Docker [7] for an ephemeral development environment, Make
[8] for command definitions, Dependabot [2] for continuous dependency
updates, GitHub Actions [9] for continuous integration, shfmt [10] for
Shell script formatting, and actionlint [11] for GitHub Actions workflow
linting.

The license for this project is the MIT license [12], matching that of
asdf. The idea being that the asdf project could relatively easily take
over maintenance of this project if they want.

The `.gitignore` file is written as an allowlist (rather than the usual
blocklist approach). Mostly just because I wanted to try it out on a
project.

In the security policy I included a "kill switch" on security support
for the initial "Early Development" timeline. The end date can be
extended if "Early Development" carries on for a while, but upon the
first stable release security support won't be provided for the v0
release line anymore.

Lastly, note that the tool versions in the `.tool-versions` file for
this project are not up-to-date as of this commit. This is intentional.
The `nightly.yml` workflow runs this project on itself, so it's a way
of testing it out in practice (eating-your-own-dog-food style).

--
1. https://asdf-vm.com/
2. https://github.com/dependabot
3. https://editorconfig.org/
4. https://github.com/hadolint/hadolint
5. https://www.shellcheck.net/
6. https://github.com/adrienverge/yamllint
7. https://docker.com
8. https://www.gnu.org/software/make/
9. https://github.com/features/actions
10. https://github.com/mvdan/sh
11. https://github.com/rhysd/actionlint
12. https://opensource.org/license/mit/