Add continuous Semgrep scan
Add a new continuous integration job to run a Semgrep scan and update
the results to GitHub's security dashboard.

The job is configured to not run for Dependabot because the secret won't
be available. The SARIF upload step is executed regardless of success or
failure to ensure it gets uploaded even if Semgrep found a problem (note
that `${{ always() }}` isn't used because that would make the step run
even if the job is cancelled, which is not what we want).