Split repo into multiple actions, add commit workflow action
Re-organize the repository to better support multiple actions for
different workflows, moving the action originally at the root into the
`/pr/action.yml`. The documentation for this action moved with it to
`/pr/README.md` - the documentation at the root is now more bare bones
with links to the available actions.

This re-organization is in order to support adding a new version of the
action, one that creates a commit instead of a Pull Request. This new
action can be found in `/commit`. It's very similar to the Pull Request
action from the user perspective, just with an additional option that
allows user to choose the branch to commit to.