Add continuous Gitleaks scan
Add a new continuous integration job to run a Gitleaks scan on commits
(to main), Pull Requests, and on a nightly schedule. To be able to run
it nightly as well as on prs and commits a new workflow was created.

The gitleaks job was is named "Secrets" to make it scanner agnostic (ie.
it could be swapped for a different secret scanning tool without it
affecting GitHub's UI). It uses a `fetch-depth` of 0 when checking out
to fetch the entire history so that all commits are scanned in nightly
jobs.

In addition, the continuous Semgrep integration was moved from the check
workflow into this workflow as it's a similar kind of check that could
benefit from being run nightly as well.