Add labels option to the Pull Request version of the action
Update the `tool-versions-update-action/pr` Action to accept as input
the option "labels" and pass that value on to the Action
`peter-evans/create-pull-request` in order to be able to add labels to
Pull Requests upon creation.

The documentation and changelog are updated accordingly. The use of this
Action by this project has also been updated to make use of this new
input.