Split audit workflow (#28)
Split the audit workflow into two separate workflows, one for secret
scanning and one for Semgrep. The primary motivator for this is the use
of a secret in the Semgrep job, which is not available in Pull Requests
from forks (and Dependabot, but this was already considered). As a
Result, the Semgrep workflow/job now only runs on the `main` branch as
well as on a schedule. The secrets job is unchanged other than that the
workflow it is a part of is now named "Secrets".