Handle missing `.tool-versions` file
Add a check to all actions in this project to detect if the project has
a `.tool-versions` file and, if not, exit early with an information
message. The goal of this change is to provide better support for users
that don't yet have `.tool-versions` file or otherwise miss it in their
project.