Pre-install tooling in development image
Update the Dockerfile development image to install development tooling
upon image creation by copying the `.tool-versions` file into the image
and running `asdf install`. This streamlines development using this
image by avoiding the need to re-install the tooling everytime the image
is used to create a ephemeral development environment.

The `dev-img` make target has been updated to trigger a rebuild when the
`.tool-versions` file changed so that the development image is always
pre-loaded with the correct tool versions.