Fixes for `plugins:` option
The sourcing of `lib/actions.sh` has been corrected based on [1], before
it did not work because the CWD is different during Action execution than
it is during local development. The current approach should work for both
(but may not locally [2]).

The invocation of the new `bin/install-plugins.sh` script has also been
fixed by, first, adding the missing `shell:` option and correcting the
quotes used in the conditinal. Second, by moving it before the "Install
asdf" step to avoid failures in that step due to a missing plugin (and
correspondingly, adding `asdf-vm/actions/setup` to make sure the asdf
CLI is installed).

Lastly, a minor improvement to the wording of the logging in the new
`bin/install-plugins.sh` script.

--
1. https://stackoverflow.com/questions/6659689
2. http://mywiki.wooledge.org/BashFAQ/028