Support configuring plugins (#36)
Update the actions of this project to all support configuring asdf
plugins in a convenient manner. For the `/commit` and `/pr` action this
is achieved through a new script (details below). For the base action
this is achieved simply by removing the step to install asdf from the
action and instead asking users to install asdf so that they can, based
on their needs, install plugins after that (but before this action is
run).

The script to install plugins is baed on the update script. It iterates
through the lines of the newly added `plugins` input and for each non-
empty line parses it as a "[NAME] [URL]" pair and invokes an adsf sub-
command to install the plugin. Lastly, there's some logic to provide the
user with information on what the outcome of this command was. Note that
this whole step is invoked only `if:` the list of plugins isn't the
empty string, this is to avoid unnecessary running the script.

To avoid duplicating the helpers previously defined in `bin/update.sh`,
those helpers are extracted into the `lib/actions.sh` file (because it
contains shared functionality specific to GitHub Actions). This file is
"imported" (`source`d) in the bin scripts so its functions can be used
there.