Use latest major version number for `uses:` in documentation
Update GitHub Actions workflows in the documentation to specify other
actions being used in terms of the latest major version number (e.g. v1)
in order to avoid having an outdated versoin of the action specified (of
course this still requires updating the docs if a new major version of
any of these actions is released, but that should be less frequent).

This also fixes a mistake in the "Full Example" workflow in the root
README.md which was missing a version number altogether for the use of
`asdf-vm/actions/install`.