Pin composite action dependencies
Update the composite action manifests that `uses:` other actions to pin
those dependencies by the commit SHA of the latest release for the
specified major version. This was done using `mheap/pin-github-action`
(v1.8.0) [1,2], which also added the "pin" comments. It uses this on
subsequent runs to update to a more recent version if available. (Unlike
Dependabot it appears that specifying the exact version (e.g. v3.0.1)
prevents the tool from updating the action to a newer version.)

The commit SHAs seen in this commit were all verified against releases
in the original Actions' repositories.

--
1. https://github.com/mheap/pin-github-action
2. https://www.npmjs.com/package/pin-github-action