Set up automation to keep composite action dependencies up-to-date
First, create a Make command that runs `pin-github-action` [1,2]
containerized in order to pin and update compose action dependencies.
To achieve this, Docker [3] is leveraged to create a ephemeral container
to which this project directory is mounted. This ephemeral container
runs `pin-update-action` (pinned to major version 1 to avoid problems
due to breaking changes in the CLI) with npx [4] on the action manifests
of this project. Notably, `npx` is configured to 1) not show the npm
update notifier in order to reduce noise, and 2) automatically accept
the installation prompt for installing `pin-github-action`.

Secondly, create a GitHub Actions workflow that is scheduled nightly
(and can be dispatched manually) to submit a Pull Request with updates
to composite action dependencies if available. To do this it simply
runs the previously mentioned Make command. The workflow's GitHub token
is utilized to avoid potential rate limiting as a result of the looking
up of versions on GitHub by `pin-github-action`.

--
1. https://github.com/mheap/pin-github-action
2. https://www.npmjs.com/package/pin-github-action
3. https://www.docker.com/
4. https://www.npmjs.com/package/npx