Downgrade `act` to working version, skip `act` updates
Undo the bump of [1] and configure the automated tooling update workflow
to stop trying to update `act`. This measure is meant to be temporary
and should be reverted when a fix for the observed problem is released
by `act`. "The observed problem" being warning messages about jobs not
being found while working just fine.

--
1. e8ed27acf10fa34038205369d5a0d8c5ac136b4b