Add output `updated-count`
Update the main script to set the 'updated-count' output initially to 0
and incrementing it by 1 every time a tool update was successful. Like
the `did-update` output, this implementation leverages the fact that
newer values override old values.

Update each Action manifest to expose this script/step output as an
Action output, as well as update the docs and changelog accordingly.