Add `skip` input (#122)
Update the main update script to accept an environment input called SKIP
which, when set, will be read line by line to look if a given *new*
version for a tool should be skipped or not. I.e., if there is a line
`$tool` `$version` such that `$version` matched the latest available
version for `$tool`, then the update won't be applied. This option is
available in all versions of the action.

Also use the `skip` input to skip problematic `act` release.