Initialize unit tests
Start a unit test suite using ShellSpec [1] as test runner. The
dependency on ShellSpec is managed through the `.tool-versions` file.
ShellSpec itself is configured through the `.shellspec` file and the
preferred way of running the unit tests is using `make test` (or `make
verify`). Note that ShellSpec is explicitly configured to run tests
using Bash, this is because the Action manifest also explicitly run the
scripts using bash (`shell: bash` directive).

In this initial commit only the `lib/actions.sh` file is covered since
it provides simple unit testable functions as a proof of concept.

--
1. https://shellspec.info/