Test the `install-plugins.sh` script
Expand the unit test suite with a test file for the `install-plugins.sh`
script covering the three main scenarios (success, plugins already
installed, and plugin could not be installed). This test uncovered a bug
where the script would exit immidiately if `asdf` errored. This is not
desired, especially in the case where a plugin is already installed. To
remedy this the `set -e` directive was removed and instead we exit
manually (still failing when any plugin couldn't be installed).

For clarity, the `spec/` directory structure has been updated. The new
structure is in accordance with the ShellSpec recommendations (current
version).

Finally, this fixes the build for the development image by explicitly
adding the ShellSpec plugin during setup.