Expand test suite for `update.sh`
Update the `update_spec.sh` test suite to have one test case for every
input option as well as one case where there's no updates and one where
there is but without any configuration.

The test for the max. capacity option uncovered a bug where `group_end`
wouldn't be called (and thus the log group not closed) when the maximum
update capacity was reached. This was fixed by breaking the loop instead
of exiting immediately.