Initialize unit tests (#128)
Start a unit test suite using ShellSpec [1] as test runner. The
dependency on ShellSpec is managed through the `.tool-versions` file.
ShellSpec itself is configured through the `.shellspec` file and the
preferred way of running the unit tests is using `make test` (or `make
verify`). Note that ShellSpec is explicitly configured to run tests
using Bash, this is because the Action manifest also explicitly run the
scripts using bash (`shell: bash` directive). It is further configured
explicitly to avoid unexpected behavior changes on updates. The spec/
directory structure is taken straight from the ShellSpec recommendation.

All source code (both lib/ and bin/) is covered by at least one test,
this was verified using the `--kcov` option [2]. Additionally, some
scenarios not seen by coverage are intentionally included (e.g. when the
`asdf` command fails).

Implementing the tests uncovered two bugs in the implementation:
1. The `install-plugins.sh` script would not handle `asdf` exit codes
   correctly, most notably failing if a plugin is already installed.
2. The `update.sh` script would not exit correctly if the maximum number
   of updates was reached, failing to close the current log group.

The labeler integration is updated to add the `test` label when specs
are updated.

--
1. https://shellspec.info/
2. https://shellspec.info/#code-coverage