Replace `Dockerfile` with `Containerfile` (#134)
Move to a `Containerfile` based ephemeral development environment
container image in order to be vendor agnostic. Nothing much should be
changed except that it has become easier to use a different container
engine when working on this project.

To achieve this, the container specification had to be updated because
the `SHELL` directive is Docker specific and not necessarily supported
by other engines. Unfortunately, just removing it doesn't work and would
result in build errors from the `asdf` installation steps. To fix this,
the installation steps have been updated based on the POSIX Shell & Git
installation instructions for asdf (since we moved from Bash to whatever
the base image's default shell is, presumably a POSIX-compliant shell).
Note that we still want to echo the sourcing of the asdf setup script to
the Bash configuration because the container entrypoint is Bash.

Dependabot does not currently support applying updates to Containerfile,
so the `docker` option is removed from the configuration. When it is
supported, it should be reconfigured (if Dependabot is still used).