Set up test coverage
Extend the unit test setup with code coverage powered by kcov [1] as
supported by ShellSpec [2]. Coverage is configured in the `.shellspec`
file and can be invoked with the `make coverage` command.

This change adds kcov as a prerequisite, though currently considered
optional because:
1. Coverage isn't considered very important, rather it's used as a
   signal for what _could_ be tested.
2. As far as I know there is currently no asdf plugin for kcov, so the
   exact version to use isn't being tracked.
3. It is also not possible to install kcov with `apk` so it isn't added
   to the development container image.

The coverage report is also continuously uploaded to Codecov [3]. The
coverage file that is uploaded is selected based on the format supported
by Codecov [4]. The overall CI configuration for this purpose is based
on ShellSpec's [2] CI configuration. This was implemented primarily to
get historical insight into coverage.

--
1. https://github.com/SimonKagstrom/kcov
2. https://github.com/shellspec/shellspec
3. https://codecov.io/
4. https://docs.codecov.com/docs/supported-report-formats