Add output `updated-tools` (#138)
Update the main script to set the 'updated-tools' output initially to an
empty string and updating it with the tool name every time a tool update
was successful. It is implemented to generate a comma separate list, w/o
any leading or trailing commas. Like the `updated-count` output, this
implementation leverages the fact that newer values override old values.

Update each Action manifest to expose this script/step output as an
Action output, as well as update the docs and changelog accordingly.

The new behaviour is also tested through the existing test suite with
updated snapshots.