Create script to automatically bump the version in `.version` (#147)
Initialize the `script/` directory dedicated to scripts used during
development. The only script currently present automates the process of
updating the `.version` file for a release based on the argument that is
provided ('major', 'minor', or 'patch'). The Release Guidelines have
been updated to prefer the use of this script when creating a release,
maintaining the manual instructions in case the script doesn't work.

The script is intentionally given the MIT-0 (aka 'no attribution')
license identifier as the intent is for it to be easier to re-use, that
is without requiring attribution upon re-use. The intend is that all
homegrown scripts will use this less restrictive license.