Drop support `act` support
Remove the use and version tracking of `act` from this project. This
decision follows unreliability of the dependency (e.g. unnecessary
updates - as seen in 0.2.55 - or confusing behavior - as seen in [1]),
as well as the fact that there is a now a unit test suite which can be
depended upon instead.

Contributors can of course still use `act` if they so choose.

--
1. e2a7b60a697c5f684d17f96126090e8760ce5212