Add outputs `updated-old-versions` and `updated-new-versions` (#167)
Update the main script to set the both the `updated-old-versions` and
`updated-new-versions` outputs initially to an empty string and updating
it with the tool's old and new version (resp.) every time a tool update
was successful. It is implemented to generate a comma separate list, w/o
any leading or trailing commas. Like previous outputs the implementation
leverages the fact that newer values override old values.

These new outputs are also available for templating in the `/commit` and
`/pr` action variants.

Update each Action manifest to expose this script/step output as an
Action output, as well as update the docs and changelog accordingly.

The new behavior is also tested through the existing test suite with
updated snapshots.