Update default `commit-message` and `pr-body` values
Change the default values for the `commit-message` input (`/commit` and
`/pr` variant) and the `pr-body` input (`/pr` variant). The aim with
these changes it provide more detailed and useful information in the
defaults. In particular for the `pr-body`, which now includes an
explanation of the behavior of the Pull Request.

The `commit-message` value is designed for a fixed-width character view.
The `pr-body` value is designed for GitHub rendered MarkDown, the use of
the `>` multiline operator ("folded") is intention as it avoids forced
newlines in the rendered content, instead letting the text flow with the
GitHub UI. (in particular, the `|` multiline operator would insert a
newline character everywhere it appears in the source.)