Create script to automatically update the `CHANGELOG.md` (#169)
Add a new script to automate the processs of updating the `CHANGELOG.md`
file for the purpose of a new release. It should be run after the
`.version` file has been updates because it is used to insert the new
version number in the changelog. The Release Guidelines have been
updated to prefer the use of this script when creating a release,
maintaining the manual instructions in case the script doesn't work.

As with the existing script, the script is given the MIT-0 license
identifier to make it easier to re-use than the source code of this
project.