Add input `rebase-strategy` to the `/pr` action variant (#187)
Add input `rebase-strategy` to the `/pr` action variant to control how
the Action rebases Pull Request.

Correspondingly, adjust the default Pull Request body to depend on the
rebase strategy. In particular adjust the explanation of the behavior of
the Pull Request at the end of the body to the behavior of the chosen
rebase strategy.