Buffer the `.tool-versions` file when updating (#259)
Update the `update.sh` script to buffer the contents of the
.tool-versions file. This is so as to avoid incorrect behavior due to
changes made to the file during the loop. For example, an update might
result in a version string that is shorter or longer than the original.
This can result in buggy behavior due to a changed offset when reading
the next line.

This was detected when an update attempting to upgrade `actionlint` from
version 1.6.27 to version 1.7.0 resulted in an error because the next
line would be started to be read at an offset of 1 (so "adolint 2.12.0"
instead of "hadolint 2.12.0").