Set expiry date for the Security Policy
Include an expiry date in the security policy primarily as a mechanism
to ensure the supported version range can reasonable be updated if no
end-of-life date has yet been set for a version.

Additionally, it nudges anyone looking at the policy to search for an
up-to-date version - which is more helpful for everyone.