Set up basic GitHub Actions SAST with ades
Integrate a (created-by-me) SAST for GitHub actions, calle ades
(<https://github.com/ericcornelissen/ades>), into the development of
this project with a Make command and CI/CD job. Running it with Docker
as a managed package manager since it's not available through asdf, as a
GitHub Action, nor as a system package.