Remove secrets check from CI (#330)
Update the CI to omit the job that checks for secrets and
instead rely solely on GitHub for secret scanning in a
continuous fashion. Individual contributors can still continue
to use any secret scanning software they like locally.

The motivation for this removal is to reduce the supply chain
surface by omitting unnecessary dependencies, given that
this dependencies adds limited value.